Rename trace_synthetic to trace_minimal
